refactor:  move monitor to separate file
- move monitor to separate file
- improve logging
